Ditch "M2Crypto" and replace it with direct calls to OpenSSL.
"M2Crypto" is still not compatible with Python 3 so I'm trying to replace it with something that is. Since we're only using one of the many features of "M2Crypto" this is quite reasonable. The signing part has been moved to this package and is based on https://pypi.python.org/pypi/poor-smime-sign/2.0.3.